Resolving inital review comments
